pyproject.toml
[tool.poetry]
name = "asyncio-book"
version = "0.1.0"
description = ""
authors = ["grm <golosov_SA@mail.ru>"]
readme = "README.md"
packages = [{include = "asyncio_book"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
aiohttp = "^3.8.5"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
